+++ libs/leon-sans-react/src/leonsans/core/constants.ts
