+++ libs/leon-sans-react/src/leonsans/core/guide.ts
