+++ libs/leon-sans-react/src/leonsans/core/paths.ts
