+++ libs/leon-sans-react/src/leonsans/core/point.ts
