+++ libs/leon-sans-react/src/leonsans/core/util.ts
