+++ libs/leon-sans-react/src/leonsans/draw/canvas/pattern.js
