+++ libs/leon-sans-react/src/leonsans/font/latin.ts
