+++ libs/leon-sans-react/src/leonsans/font/number.ts
