+++ libs/leon-sans-react/src/package.json
